pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ship-hull-agent"
version = "0.2.0"
description = "基于 LangChain + FAISS + YOLO + Qwen3.5 VLM 的智能船弦号识别系统"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",

    "openai>=1.0",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "rich>=13.0",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "ultralytics>=8.0",
    # Web 服务
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
ship-hull = "cli:main"

[tool.setuptools.packages.find]
include = ["agent*", "database*", "tools*", "cli*", "pipeline*", "web*"]
